refactor(formValidationLogic): extract form response validation into pure logic
